pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "open-finance-pydanticai"
version = "0.1.0"
description = "Open Finance API using PydanticAI"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pydantic-ai>=1.18.0",
    "openai>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.25.0",
    "pydantic-settings>=2.0.0",
    "numpy>=1.24.0",
    "numpy-financial>=1.0.0",
    "pandas>=2.0.0",
    "scipy>=1.10.0",
    "logfire>=0.0.1",
    "gradio>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
]
quant = [
    "QuantLib>=1.31",
]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

